intel_adsp: cavs: add gdb support
Uses SRAM debug region to pass GDB messages to host, which is exposed
via PTY by cavstool with -d arg. GDBSTUB_ENTER_IMMEDIATELY was added so
that entering the stub pre-main could be disabled, which is required for
kernel driver as the boot process needs to complete without
interruption. Cache eviction had to be added to Zephyr's stub
implementation to support breakpoints properly.

+++ include/zephyr/arch/xtensa/gdbstub.h
@@ -39,7 +39,7 @@ struct xtensa_register {
 	uint32_t	val;
 
 	/* GDB register index (for p/P packets) */
-	uint8_t		idx;
+	uint32_t	idx;
 
 	/* Size of register */
 	uint8_t		byte_size;
